security 인증/인가 service 단에 적용

+++ src/main/java/com/example/hh99miniproject8/service/CommentService.java
@@ -8,10 +8,6 @@ import com.example.hh99miniproject8.entity.User;
 import com.example.hh99miniproject8.exception.CustomException;
 import com.example.hh99miniproject8.repository.CommentRepository;
 import com.example.hh99miniproject8.repository.PostRepository;
-import com.example.hh99miniproject8.repository.UserRepository;
-import com.example.hh99miniproject8.security.jwt.JwtProvider;
-import io.jsonwebtoken.Claims;
-import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -19,72 +15,56 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import static com.example.hh99miniproject8.exception.ErrorCode.*;
-import static com.example.hh99miniproject8.exception.ErrorCode.USER_NOT_FOUND;
 
 @Service
 @RequiredArgsConstructor
 public class CommentService {
     private final CommentRepository commentRepository;
     private final PostRepository postRepository;
-    private final UserRepository userRepository;
-    private final JwtProvider jwtProvider;
 
+    //========================API 메서드==========================
     //댓글 생성 API
     @Transactional
-    public ResponseEntity<CommentResponseDto> createComment(Long id, CommentRequestDto requestDTO, HttpServletRequest httpServletRequest) {
-        User user = tokenCheck(httpServletRequest);
-        Post post = checkPost(id);
-        Comment comment = commentRepository.save(new Comment(requestDTO, user, post));
+    public ResponseEntity<CommentResponseDto> createComment(Long id, CommentRequestDto requestDTO, User user) {
+        Post post = isPostExist(id);
+        Comment comment = commentRepository.saveAndFlush(new Comment(requestDTO, user, post));
         return ResponseEntity.status(HttpStatus.CREATED).body(new CommentResponseDto(comment));
-
     }
 
     //댓글 수정 API
     @Transactional
-    public ResponseEntity<CommentResponseDto> updateComment(Long id, CommentRequestDto requestDTO, HttpServletRequest httpServletRequest) {
-        User user = tokenCheck(httpServletRequest);
-        Comment comment = checkComment(id);
-        comment.update(requestDTO, user);
+    public ResponseEntity<CommentResponseDto> updateComment(Long id, CommentRequestDto requestDTO, User user) {
+        // 게시글 존재유무 체크
+        Comment comment = isCommentExist(id);
+        // 작성자와, 유저 매치체크
+        if(!checkAuthorIdMatch(comment, user)) {
+            throw new CustomException(WRITER_ONLY_MODIFY);
+        }
+        comment.update(requestDTO);
         return ResponseEntity.status(HttpStatus.OK).body(new CommentResponseDto(comment));
-
     }
 
     //댓글 삭제 API
     @Transactional
-    public ResponseEntity<String> deleteComment(Long id, HttpServletRequest httpServletRequest) {
-        User user = tokenCheck(httpServletRequest);
-        Comment comment = checkComment(id);
+    public ResponseEntity<String> deleteComment(Long id, User user) {
+        // 게시글 존재유무 체크
+        Comment comment = isCommentExist(id);
+        // 작성자와, 유저 매치체크
+        if(!checkAuthorIdMatch(comment, user))
+            throw new CustomException(WRITER_ONLY_DELETE);
+
         commentRepository.delete(comment);
         return ResponseEntity.status(HttpStatus.OK).body("댓글 식제 성공");
-
     }
+    //========================API 메서드==========================
 
 
-    // 토큰 검사
-    public User tokenCheck(HttpServletRequest httpServletRequest) {
-        String token = jwtProvider.resolveAccessToken(httpServletRequest);
-        Claims claims;
-
-        if (token != null) {
-            if (jwtProvider.validateToken(token)) {
-                // 토큰에서 사용자 정보 가져오기
-                claims = jwtProvider.getUserInfoFromToken(token);
-            } else {
-                throw new CustomException(TOKEN_NOT_FOUND);
-            }
-
-            // 토큰에서 가져온 사용자 정보를 사용하여 DB 조회
-            User user = userRepository.findByUsername(claims.getSubject()).orElseThrow(
-                    () -> new CustomException(USER_NOT_FOUND)
-            );
-            return user;
-
-        }
-        return null;
-    }
-
+    //========================외부 메서드==========================
     // 댓글 존재 여부 확인
-    private Comment checkComment(Long id){
+    // chore 석빈
+    // check는 여러가지 의미를 가진 단어이고 메서드에서 수행하는 동작을 직관적으로 알기 힘들기에
+    // checkComment -> isCommentExist 로 변경했습니다.
+    private Comment isCommentExist(Long id){
         Comment comment = commentRepository.findById(id).orElseThrow(
                 () -> new CustomException(COMMENT_NOT_FOUND)
         );
@@ -92,10 +72,17 @@ public class CommentService {
     }
 
     // 게시글 존재 여부 확인
-    private Post checkPost(Long id){
-        Post post = postRepository.findById(id).orElseThrow(
+    private Post isPostExist(Long id){
+        return postRepository.findById(id).orElseThrow(
                 () -> new CustomException(POST_NOT_FOUND)
         );
-        return post;
     }
+
+    // 게시글과 게시글을 변경하려는 요청을 보낸 유저가 일치하는지 여부 체크 default return value : true
+    private boolean checkAuthorIdMatch(Comment comment, User user){
+        if(!comment.getUser().getUsername().equals(user.getUsername()))
+            return false;
+        return true;
+    }
+    //========================외부 메서드==========================
 }
